feat: implement Plan 5 - Router Split
Split 942-line metric.ts into focused routers:

New routers:
- goal.ts: Goal CRUD (get, upsert, delete)
- manual-metric.ts: Manual metric operations (create, addDataPoints, updateChart, getForUser, getById)
- pipeline.ts: Fire-and-forget pipeline (refresh, regenerate, getProgress)

Removed:
- transformer.ts (permanently deleted, procedures moved to pipeline.ts and manual-metric.ts)

Key changes:
- Implement fire-and-forget pattern for pipeline operations
- Frontend now polls pipeline.getProgress for status updates
- Updated all frontend components to use new router paths
- metric.ts reduced to CRUD + integration data fetching only
