+++ src/server/api/routers/metric.ts
@@ -2,16 +2,7 @@ import type { PrismaClient } from "@prisma/client";
 import { TRPCError } from "@trpc/server";
 import { z } from "zod";
 
-import {
-  type ChartDataForGoal,
-  type GoalInput,
-  calculateGoalProgress,
-} from "@/lib/goals";
 import { getTemplate } from "@/lib/integrations";
-import type {
-  ChartConfig,
-  ChartTransformResult,
-} from "@/lib/metrics/transformer-types";
 import { fetchData } from "@/server/api/services/data-fetching";
 import {
   createChartTransformer,
@@ -379,519 +370,6 @@ export const metricRouter = createTRPCRouter({
       await invalidateCacheByTags(ctx.db, cacheTags);
 
       return { success: true };
-    }),
-
-  // ===========================================================================
-  // Goal Management
-  // ===========================================================================
-
-  getGoal: workspaceProcedure
-    .input(z.object({ metricId: z.string() }))
-    .query(async ({ ctx, input }) => {
-      // Verify metric belongs to org
-      const metric = await getMetricAndVerifyAccess(
-        ctx.db,
-        input.metricId,
-        ctx.workspace.organizationId,
-      );
-
-      const goal = await ctx.db.metricGoal.findUnique({
-        where: { metricId: input.metricId },
-      });
-
-      // Get the chart's cadence and chartConfig for goal calculation
-      const dashboardChart = await ctx.db.dashboardChart.findFirst({
-        where: { metricId: input.metricId },
-        select: {
-          chartConfig: true,
-          chartTransformer: { select: { cadence: true } },
-          metric: { select: { integrationId: true, endpointConfig: true } },
-        },
-      });
-
-      // Manual metrics store cadence in endpointConfig (lowercase), convert to uppercase Prisma Cadence
-      const isManualMetric = dashboardChart?.metric?.integrationId === null;
-      const rawManualCadence = (
-        dashboardChart?.metric?.endpointConfig as { cadence?: string } | null
-      )?.cadence;
-      const manualCadence = rawManualCadence
-        ? (rawManualCadence.toUpperCase() as "DAILY" | "WEEKLY" | "MONTHLY")
-        : null;
-      const cadence = isManualMetric
-        ? manualCadence
-        : (dashboardChart?.chartTransformer?.cadence ?? null);
-
-      // Parse chartConfig for unified metadata access
-      const chartConfig = dashboardChart?.chartConfig as unknown as ChartConfig;
-      const chartConfigResult =
-        dashboardChart?.chartConfig as unknown as ChartTransformResult;
-
-      // Get valueLabel from chartConfig (ChartTransformer output) as primary source
-      // Fall back to DataIngestionTransformer for backward compatibility
-      let valueLabel: string | null = null;
-      if (chartConfigResult?.valueLabelOverride) {
-        valueLabel = chartConfigResult.valueLabelOverride;
-      } else if (chartConfigResult?.valueLabel) {
-        valueLabel = chartConfigResult.valueLabel;
-      } else {
-        // Fallback: try DataIngestionTransformer (keyed by metricId for independent metrics)
-        const transformer = await ctx.db.dataIngestionTransformer.findUnique({
-          where: { templateId: metric.id },
-          select: { valueLabel: true },
-        });
-        valueLabel = transformer?.valueLabel ?? null;
-      }
-
-      // Extract current value from chartConfig
-      let currentValue: number | null = null;
-      let currentValueLabel: string | null = null;
-      if (chartConfig?.chartData && chartConfig.chartData.length > 0) {
-        const latestData =
-          chartConfig.chartData[chartConfig.chartData.length - 1];
-        const primaryKey = chartConfig.dataKeys?.[0];
-        if (latestData && primaryKey) {
-          const val = latestData[primaryKey];
-          if (typeof val === "number") {
-            currentValue = val;
-            // Use chartConfig label for what's being displayed, fallback to valueLabel
-            currentValueLabel =
-              chartConfig.chartConfig?.[primaryKey]?.label ??
-              valueLabel ??
-              primaryKey;
-          }
-        }
-      }
-
-      // If no goal, return current value info for goal creation context
-      if (!goal) {
-        return {
-          goal: null,
-          progress: null,
-          cadence,
-          currentValue,
-          currentValueLabel,
-          valueLabel,
-        };
-      }
-
-      // If no chart or no cadence, return goal without progress
-      if (!cadence) {
-        return {
-          goal,
-          progress: null,
-          cadence: null,
-          currentValue,
-          currentValueLabel,
-          valueLabel,
-        };
-      }
-
-      // Convert MetricGoal to GoalInput
-      const goalInput: GoalInput = {
-        goalType: goal.goalType,
-        targetValue: goal.targetValue,
-        baselineValue: goal.baselineValue,
-        baselineTimestamp: goal.baselineTimestamp,
-        onTrackThreshold: goal.onTrackThreshold,
-      };
-
-      // Convert ChartConfig to ChartDataForGoal
-      const chartDataForGoal: ChartDataForGoal = {
-        chartData: chartConfig?.chartData ?? [],
-        xAxisKey: chartConfig?.xAxisKey ?? "date",
-        dataKeys: chartConfig?.dataKeys ?? [],
-      };
-
-      const progress = calculateGoalProgress(
-        goalInput,
-        cadence,
-        chartDataForGoal,
-      );
-      return {
-        goal,
-        progress,
-        cadence,
-        currentValue,
-        currentValueLabel,
-        valueLabel,
-      };
-    }),
-
-  upsertGoal: workspaceProcedure
-    .input(
-      z.object({
-        metricId: z.string(),
-        goalType: z.enum(["ABSOLUTE", "RELATIVE"]),
-        targetValue: z.number().positive(),
-        onTrackThreshold: z.number().min(0).max(100).optional(),
-      }),
-    )
-    .mutation(async ({ ctx, input }) => {
-      // Verify metric belongs to org
-      const metric = await getMetricAndVerifyAccess(
-        ctx.db,
-        input.metricId,
-        ctx.workspace.organizationId,
-      );
-
-      // Check if goal already exists
-      const existingGoal = await ctx.db.metricGoal.findUnique({
-        where: { metricId: input.metricId },
-      });
-
-      // Capture baseline only on goal creation (not update)
-      let baselineValue: number | null = null;
-      let baselineTimestamp: Date | null = null;
-
-      if (!existingGoal) {
-        // NEW GOAL: Capture baseline from current chart data
-        const dashboardChart = await ctx.db.dashboardChart.findFirst({
-          where: { metricId: input.metricId },
-          select: { chartConfig: true },
-        });
-
-        const chartConfig = dashboardChart?.chartConfig as ChartConfig | null;
-        if (chartConfig?.chartData?.length) {
-          const latestData =
-            chartConfig.chartData[chartConfig.chartData.length - 1];
-          const primaryKey = chartConfig.dataKeys?.[0];
-          if (
-            latestData &&
-            primaryKey &&
-            typeof latestData[primaryKey] === "number"
-          ) {
-            baselineValue = latestData[primaryKey];
-            baselineTimestamp = new Date();
-          }
-        }
-      }
-
-      const goal = await ctx.db.metricGoal.upsert({
-        where: { metricId: input.metricId },
-        create: {
-          metricId: input.metricId,
-          goalType: input.goalType,
-          targetValue: input.targetValue,
-          baselineValue,
-          baselineTimestamp,
-          onTrackThreshold: input.onTrackThreshold,
-        },
-        update: {
-          goalType: input.goalType,
-          targetValue: input.targetValue,
-          onTrackThreshold: input.onTrackThreshold,
-        },
-      });
-
-      // Invalidate dashboard cache
-      const cacheTags = [`dashboard_org_${ctx.workspace.organizationId}`];
-      if (metric.teamId) {
-        cacheTags.push(`dashboard_team_${metric.teamId}`);
-      }
-      await invalidateCacheByTags(ctx.db, cacheTags);
-
-      return goal;
-    }),
-
-  deleteGoal: workspaceProcedure
-    .input(z.object({ metricId: z.string() }))
-    .mutation(async ({ ctx, input }) => {
-      // Verify metric belongs to org
-      const metric = await getMetricAndVerifyAccess(
-        ctx.db,
-        input.metricId,
-        ctx.workspace.organizationId,
-      );
-
-      await ctx.db.metricGoal.delete({
-        where: { metricId: input.metricId },
-      });
-
-      // Invalidate dashboard cache
-      const cacheTags = [`dashboard_org_${ctx.workspace.organizationId}`];
-      if (metric.teamId) {
-        cacheTags.push(`dashboard_team_${metric.teamId}`);
-      }
-      await invalidateCacheByTags(ctx.db, cacheTags);
-
-      return { success: true };
-    }),
-
-  // ===========================================================================
-  // Manual Metrics
-  // ===========================================================================
-
-  /**
-   * Create a manual metric (user-entered KPI without integration)
-   * No data fetching or transformer generation needed at creation time.
-   */
-  createManual: workspaceProcedure
-    .input(
-      z.object({
-        name: z.string().min(1).max(100),
-        description: z.string().optional(),
-        unitType: z.enum(["number", "percentage"]),
-        cadence: z.enum(["daily", "weekly", "monthly"]),
-        teamId: z.string(),
-      }),
-    )
-    .mutation(async ({ ctx, input }) => {
-      // Get position for dashboard chart
-      const count = await ctx.db.dashboardChart.count({
-        where: { organizationId: ctx.workspace.organizationId },
-      });
-
-      // Create metric + dashboard chart in transaction
-      const dashboardChart = await ctx.db.$transaction(
-        async (tx) => {
-          const metric = await tx.metric.create({
-            data: {
-              name: input.name,
-              description: input.description,
-              organizationId: ctx.workspace.organizationId,
-              teamId: input.teamId,
-              // Manual metrics have no integration or template
-              integrationId: null,
-              templateId: null,
-              // Store manual config in endpointConfig
-              endpointConfig: {
-                type: "manual",
-                unitType: input.unitType,
-                cadence: input.cadence,
-              },
-              pollFrequency: "manual",
-              nextPollAt: null,
-            },
-          });
-
-          return tx.dashboardChart.create({
-            data: {
-              organizationId: ctx.workspace.organizationId,
-              metricId: metric.id,
-              chartType: "line",
-              chartConfig: {},
-              size: "medium",
-              position: count,
-            },
-            include: {
-              metric: {
-                include: {
-                  integration: true,
-                  roles: true,
-                },
-              },
-            },
-          });
-        },
-        { timeout: 15000 },
-      );
-
-      // Invalidate Prisma cache for dashboard queries
-      const cacheTags = [`dashboard_org_${ctx.workspace.organizationId}`];
-      cacheTags.push(`dashboard_team_${input.teamId}`);
-      await invalidateCacheByTags(ctx.db, cacheTags);
-
-      return dashboardChart;
-    }),
-
-  /**
-   * Add or update data points for a manual metric
-   * Uses upsert to handle both new entries and updates
-   */
-  addDataPoints: workspaceProcedure
-    .input(
-      z.object({
-        metricId: z.string(),
-        dataPoints: z.array(
-          z.object({
-            timestamp: z.coerce.date(),
-            value: z.number(),
-          }),
-        ),
-      }),
-    )
-    .mutation(async ({ ctx, input }) => {
-      // Verify metric exists and belongs to user's organization
-      const metric = await ctx.db.metric.findFirst({
-        where: {
-          id: input.metricId,
-          organizationId: ctx.workspace.organizationId,
-        },
-      });
-
-      if (!metric) {
-        throw new TRPCError({
-          code: "NOT_FOUND",
-          message: "Metric not found",
-        });
-      }
-
-      // Verify this is a manual metric
-      if (metric.integrationId !== null) {
-        throw new TRPCError({
-          code: "BAD_REQUEST",
-          message: "Can only add data points to manual metrics",
-        });
-      }
-
-      // Batch upsert data points
-      const results = await Promise.all(
-        input.dataPoints.map((dp) =>
-          ctx.db.metricDataPoint.upsert({
-            where: {
-              metricId_timestamp: {
-                metricId: input.metricId,
-                timestamp: dp.timestamp,
-              },
-            },
-            create: {
-              metricId: input.metricId,
-              timestamp: dp.timestamp,
-              value: dp.value,
-            },
-            update: {
-              value: dp.value,
-            },
-          }),
-        ),
-      );
-
-      await ctx.db.metric.update({
-        where: { id: input.metricId },
-        data: { lastFetchedAt: new Date() },
-      });
-
-      return {
-        success: true,
-        savedCount: results.length,
-      };
-    }),
-
-  /**
-   * Get all manual metrics assigned to a user via their roles
-   * Groups metrics by cadence for the check-in page
-   */
-  getManualMetricsForUser: workspaceProcedure
-    .input(z.object({ userId: z.string() }))
-    .query(async ({ ctx, input }) => {
-      // Find all roles assigned to this user that have manual metrics
-      const rolesWithMetrics = await ctx.db.role.findMany({
-        where: {
-          assignedUserId: input.userId,
-          metricId: { not: null },
-          metric: {
-            organizationId: ctx.workspace.organizationId,
-            integrationId: null, // Manual metrics only
-          },
-        },
-        include: {
-          metric: {
-            include: {
-              dataPoints: {
-                orderBy: { timestamp: "desc" },
-                take: 10,
-              },
-              dashboardCharts: {
-                select: { id: true },
-                take: 1,
-              },
-            },
-          },
-          team: {
-            select: { id: true, name: true },
-          },
-        },
-      });
-
-      // Group by cadence
-      type MetricWithRole = {
-        metric: NonNullable<(typeof rolesWithMetrics)[0]["metric"]>;
-        role: {
-          id: string;
-          title: string;
-          team: { id: string; name: string };
-        };
-      };
-
-      const grouped: {
-        daily: MetricWithRole[];
-        weekly: MetricWithRole[];
-        monthly: MetricWithRole[];
-      } = {
-        daily: [],
-        weekly: [],
-        monthly: [],
-      };
-
-      for (const role of rolesWithMetrics) {
-        if (!role.metric) continue;
-
-        const config = role.metric.endpointConfig as {
-          type?: string;
-          cadence?: string;
-        } | null;
-        const cadence = config?.cadence ?? "weekly";
-
-        const metricWithRole: MetricWithRole = {
-          metric: role.metric,
-          role: {
-            id: role.id,
-            title: role.title,
-            team: role.team,
-          },
-        };
-
-        if (cadence === "daily") {
-          grouped.daily.push(metricWithRole);
-        } else if (cadence === "monthly") {
-          grouped.monthly.push(metricWithRole);
-        } else {
-          grouped.weekly.push(metricWithRole);
-        }
-      }
-
-      return grouped;
-    }),
-
-  /**
-   * Get a single manual metric by ID for the metric check-in page
-   */
-  getManualMetricById: workspaceProcedure
-    .input(z.object({ metricId: z.string() }))
-    .query(async ({ ctx, input }) => {
-      const metric = await ctx.db.metric.findFirst({
-        where: {
-          id: input.metricId,
-          organizationId: ctx.workspace.organizationId,
-          integrationId: null, // Manual metrics only
-        },
-        include: {
-          dataPoints: {
-            orderBy: { timestamp: "desc" },
-            take: 10,
-          },
-          roles: {
-            include: {
-              team: {
-                select: { id: true, name: true },
-              },
-            },
-          },
-          dashboardCharts: {
-            select: { id: true },
-            take: 1,
-          },
-        },
-      });
-
-      if (!metric) {
-        throw new TRPCError({
-          code: "NOT_FOUND",
-          message: "Manual metric not found",
-        });
-      }
-
-      return metric;
     }),
 
   // ===========================================================================
